requirements.txt
django
django-cors-headers
django-ninja
django-ninja-jwt[crypto]
dj-database-url
psycopg[binary] # psycopg
pydantic[email]
python-decouple
rav
faker